Move validation cases to examples/legacy_validation/ (#87)
Relocate the legacy validation cases (case1-4) into
examples/legacy_validation/ as history-preserving renames, and update the
corresponding import in tests/test_examples.py. The file contents are
otherwise unchanged.
